+++ database/migrations/2025_11_20_233948_create_cart_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
 
             $table->integer('qty')->default(1);
-            $table->decimal('price', 12, 2); // snapshot harga saat ditambah
+            $table->unsignedBigInteger('price');
 
             $table->timestamps();
         });
